Переносит отрисовку больших фото в соответствующий блок
