+++ js/main.js
@@ -10,12 +10,23 @@ import {closeForm} from './user-form.js';
 import {getData} from './api.js';
 import {showImageFilters} from './filter.js';
 
-getData((photos) => {
+// getData((photos) => {
+//   renderPhotos(photos);
+//   console.log(photos);
+//   const filteredPhotos = photos.slice(0, 10);
+//   console.log(filteredPhotos);
+//   setTimeout(showImageFilters, 1000);
+// });
+
+const onSuccess = function (photos) {
   renderPhotos(photos);
+  // вызов filters ();
   console.log(photos);
   const filteredPhotos = photos.slice(0, 10);
   console.log(filteredPhotos);
-  setTimeout(showImageFilters, 1000);
-});
+  showImageFilters();
+};
+
+getData(onSuccess);
 
 setUserFormSubmit(closeForm);
